refactor seprate Rank and Statistic
