+++ src/main/java/lotto/Rank.java
@@ -1,7 +1,6 @@
 package lotto;
 
 import java.util.Arrays;
-import java.util.List;
 
 public enum Rank {
 
@@ -26,19 +25,6 @@ public enum Rank {
                 .filter(rank -> rank.matchCount == matchCount)
                 .findFirst()
                 .orElseThrow(() -> new IllegalArgumentException("0 ~ 6이외의 matchCount 는 발생할 수 없습니다."));
-    }
-
-    public static double calculateRate(int price, List<Rank> ranks) {
-        int totalWinnings = calculateTotalWinnings(ranks);
-
-        double rate = (double) totalWinnings / price;
-        return Math.floor(rate * 100) / 100;
-    }
-
-    private static int calculateTotalWinnings(List<Rank> ranks) {
-        return ranks.stream()
-                .mapToInt(i -> i.winnings)
-                .sum();
     }
 
     public int getMatchCount() {
